build: init 2 modules
shared is a KMP module to contain schemas between the frontend and backend
server is the main module for the ktor server
