+++ .gitignore
@@ -23,3 +23,4 @@ gradle-app.setting
 
 # Ignore Gradle build output directory
 build
+/.kotlin/
