--- app/build.gradle.kts
@@ -1,17 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    id("buildlogic.kotlin-application-conventions")
-}
-
-dependencies {
-    implementation("org.apache.commons:commons-text")
-    implementation(project(":utilities"))
-}
-
-application {
-    // Define the main class for the application.
-    mainClass = "org.example.app.AppKt"
-}
