--- app/src/main/kotlin/org/iskdevelop/mithra/App.kt
@@ -1,14 +1,0 @@
-/*
- * This source file was generated by the Gradle 'init' task
- */
-package org.iskdevelop.mithra
-
-import org.example.utilities.StringUtils
-
-import org.apache.commons.text.WordUtils
-
-fun main() {
-    val tokens = StringUtils.split(MessageUtils.getMessage())
-    val result = StringUtils.join(tokens)
-    println(WordUtils.capitalize(result))
-}
