--- app/src/main/kotlin/org/iskdevelop/mithra/MessageUtils.kt
@@ -1,10 +1,0 @@
-/*
- * This source file was generated by the Gradle 'init' task
- */
-package org.iskdevelop.mithra
-
-class MessageUtils {
-    companion object {
-        fun getMessage(): String = "Hello      World!"
-    }
-}
