+++ buildSrc/src/main/kotlin/JvmToolChain.kt
@@ -1,0 +1,5 @@
+import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
+
+fun KotlinProjectExtension.configureJvmToolChain() {
+    jvmToolchain(21)
+}
