--- buildSrc/src/main/kotlin/buildlogic.kotlin-application-conventions.gradle.kts
@@ -1,11 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    // Apply the common convention plugin for shared build configuration between library and application projects.
-    id("buildlogic.kotlin-common-conventions")
-
-    // Apply the application plugin to add support for building a CLI application in Java.
-    application
-}
