+++ buildSrc/src/main/kotlin/kotlin-jvm-base.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    alias(libs.plugins.kotlin.jvm)
+    alias(libs.plugins.kotlin.serialization)
+    alias(libs.plugins.shadow)
+}
+
+kotlin {
+    configureJvmToolChain()
+}
