+++ buildSrc/src/main/kotlin/kotlin-multiplatform-base.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    alias(libs.plugins.kotlin.multiplatform)
+    alias(libs.plugins.kotlin.serialization)
+}
+
+kotlin {
+    configureJvmToolChain()
+}
