+++ gradle/libs.versions.toml
@@ -2,7 +2,20 @@
 # https://docs.gradle.org/current/userguide/platforms.html#sub::toml-dependencies-format
 
 [versions]
-kotlin-gradle-plugin = "2.0.0"
+kotlin = "2.1.10"
+shadow = "9.0.0-beta11"
+graalvm-native-image = "0.10.6"
+
+kotlinx-serialization = "1.7.3"
+kotlinx-coroutines = "1.8.1"
+ktor = "3.0.0"
 
 [libraries]
-kotlin-gradle-plugin = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", version.ref = "kotlin-gradle-plugin" }
+kotlin-gradle-plugin = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", version.ref = "kotlin" }
+
+[plugins]
+kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
+kotlin-multiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
+kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
+shadow = { id = "com.gradleup.shadow", version.ref = "shadow" }
+graalvm-native-image = { id = "org.graalvm.buildtools.native", version.ref = "graalvm-native-image" }
