--- list/build.gradle.kts
@@ -1,7 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    id("buildlogic.kotlin-library-conventions")
-}
