+++ server/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    id("kotlin-jvm-base")
+    application
+}
+
+application {
+    mainClass = "org.iskdevelop.mithra.server.AppKt"
+}
