+++ server/src/main/kotlin/org/iskdevelop/mithra/server/App.kt
@@ -1,0 +1,9 @@
+/*
+ * This source file was generated by the Gradle 'init' task
+ */
+package org.iskdevelop.mithra.server
+
+
+fun main() {
+
+}
