+++ shared/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    id("kotlin-multiplatform-base")
+}
+
+kotlin {
+    jvm()
+    js(IR) {
+        generateTypeScriptDefinitions()
+        browser()
+    }
+
+    sourceSets {
+        commonMain {
+            dependencies {
+
+            }
+        }
+    }
+}
