--- utilities/build.gradle.kts
@@ -1,11 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-plugins {
-    id("buildlogic.kotlin-library-conventions")
-}
-
-dependencies {
-    api(project(":list"))
-}
