--- utilities/src/main/kotlin/org/example/utilities/JoinUtils.kt
@@ -1,22 +1,0 @@
-/*
- * This source file was generated by the Gradle 'init' task
- */
-package org.example.utilities
-
-import org.example.list.LinkedList
-
-class JoinUtils {
-    companion object {
-        fun join(source: LinkedList): String {
-            val result = StringBuilder()
-            for (i in 0 until source.size()) {
-                if (result.isNotEmpty()) {
-                    result.append(" ")
-                }
-                result.append(source.get(i))
-            }
-
-            return result.toString()
-        }
-    }
-}
