--- utilities/src/main/kotlin/org/example/utilities/StringUtils.kt
@@ -1,18 +1,0 @@
-/*
- * This source file was generated by the Gradle 'init' task
- */
-package org.example.utilities
-
-import org.example.list.LinkedList
-
-class StringUtils {
-    companion object {
-        fun join(source: LinkedList): String {
-            return JoinUtils.join(source)
-        }
-
-        fun split(source: String): LinkedList {
-            return SplitUtils.split(source)
-        }
-    }
-}
